Update develop_test_and_review.md
See the clarification on feedback and using the course groups. I find it preferable to create course groups for the semester and adjust them after The Ritual Project Brainstorm when it is clear which students are not participating. This allows the course groups to get in a groove and become accustom to communicating with one another. It alleviates student anxiety for the semester. It also ensures faculty don't miss assigning something like "prototype feedback"  or new groups for a project and allows them to focus on substantive feedback and students without adding more tasks for faculty in the class. Students in groups can be accountable for a task like this and it is a great job for the team leader to take over.